Adds the ability to ignore duplicate notices
This PR adds an option on the application edit page which allows apps to
ignore additional notifications on the same error. This is to facilitate
noisy apps that can potentially spam errbit. This will ensure that
whilst all errors are still saved we only write one notice. The default
behaviour will remain the same for existing applications this feature
will need enabled on the applications that wish to use it.